requirements.txt
pandas
Pillow
plotly
kaleido==0.1.0.post1